Get rid of unnecessary dependencies
The time dependency can be replaced with the std equivalent, except for
one location where the last modified date is parsed as a date, and then
stored as the number of seconds since 1970-01-01. This seems like a
strange representation of a date anyway, and not worth bringing in a
dependency for.

The rustc-serialize dependency does not seem to be used at all. If
library users needs (de)serilization they can add it themselves, and
they are most likely using serde anyway as rustc-serialize is
deprecated.