Avoid `use Enum::*`
The `use Enum::*` for enums in `match` blocks to avoid repeating
the enum name can be a footgun in case a variant is later removed -
inside the `match` block the removed variant becomes a catchall variant.

And if one variant is removed and another is added, the removed variant
will catch the newly added one, without us needing to handle the newly
added variant.

For a better explanation: https://youtu.be/8j_FbjiowvE?t=97